change exception, and move test.
