+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/transaction/buffer/TransactionLowWaterMarkTest.java
@@ -34,9 +34,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.bookkeeper.mledger.impl.PositionImpl;
 import org.apache.commons.collections4.map.LinkedMap;
 import org.apache.pulsar.broker.service.BrokerService;
+import org.apache.pulsar.broker.service.BrokerServiceException;
 import org.apache.pulsar.broker.service.Topic;
 import org.apache.pulsar.broker.service.persistent.PersistentSubscription;
+import org.apache.pulsar.broker.service.persistent.PersistentTopic;
 import org.apache.pulsar.broker.transaction.TransactionTestBase;
+import org.apache.pulsar.broker.transaction.buffer.impl.TopicTransactionBuffer;
 import org.apache.pulsar.broker.transaction.pendingack.impl.PendingAckHandleImpl;
 import org.apache.pulsar.client.api.Consumer;
 import org.apache.pulsar.client.api.Message;
@@ -283,6 +286,23 @@ public class TransactionLowWaterMarkTest extends TransactionTestBase {
 
         } else {
             fail();
+        }
+    }
+
+    @Test
+    public void testTBLowWaterMark() throws Exception {
+        PersistentTopic persistentTopic = (PersistentTopic) getPulsarServiceList().get(0)
+                .getBrokerService()
+                .getTopic(NAMESPACE1 + "/test", true)
+                .get().get();
+        TopicTransactionBuffer topicTransactionBuffer = new TopicTransactionBuffer(persistentTopic);
+        TxnID txnID = new TxnID(1, 1);
+        topicTransactionBuffer.commitTxn(txnID, txnID.getLeastSigBits()).get();
+        try {
+            topicTransactionBuffer.appendBufferToTxn(txnID, 0L, null).get();
+            fail();
+        } catch (Exception e) {
+            Assert.assertTrue(e.getCause() instanceof BrokerServiceException.NotAllowedException);
         }
     }
 }
