+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/ClientCnx.java
@@ -692,7 +692,7 @@ public class ClientCnx extends PulsarHandler {
             producers.get(producerId).terminated(this);
             break;
         case NotAllowedError:
-            producers.get(producerId).recoverNotAllowedError(sequenceId);
+            producers.get(producerId).recoverNotAllowedError(sequenceId, sendError.getMessage());
             break;
 
         default:
